requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
aiofiles==23.2.1
fastapi-chameleon-0.1.14